30
Internal milestone sequence is pretty much complete.  Still need to adjust the ADD modal to be able to remove added fields.  Once that is figured out I will move to the external team section.    Also just realized that I need to include the CLIENT NAME in the base J info display.  Also adjusted the table relations to delete on cascade for EMPLOYEE delete.